rfb: Update incomplete handling in parser.
This commit adds an updated incomplete handling for the RFB-Parser. If incomplete data is processed, the successfully consumed
position and length of remainder + 1 is returned. If the next packet is not empty suricata will call the parser again.

This commit is a result of discussion on https://github.com/OISF/suricata/pull/4792.