output/json: add new file type of plugin
The plugin file type is a dynamically registered file type
that can be used by a plugin to register a custom file
type.

A plugin file type is not much different than the built-in
ones, the plugin is passed a LogFileCtx structure that it
can fill out, in particular the Write, and optionally
the Close callbacks.

TODO: May want to consider using a smaller struct for
  plugin outputs. LogFileCtx contains a lot of internal
  details the a plugin output type may not need.