xor: implement xor and xor_data keywords
Adds support for the xor keyword to decode payload or sticky buffer
data given an xor key. This will place the decoded data into xor_data's
sticky buffer. The xor_data keyword will then activate the sticky
buffer to perform matches on the decoded data.

The implementation is based off of the base64 decoding keywords except
an inspection buffer is used to store decoded bytes instead of
hard-coding values in the DetectEngineThreadCtx.

Utility functions for decoding hex and xor data were implemented to
support this feature.