suricata: indicate that Suricata is ready for packets.
Depending on the configuration used, the time it takes for
Suricata to initialize and have worker threads ready to
process packets can vary dramatically.

In order to facility quicker end-to-end/integration
tests, having an indication for when this state has
been reached would be beneficial.

This change first waits for workers to unpause, and then
waits for them to actively enter their packet polling loops
before outputting an log string which test frameworks
can use to determine that Suricata is ready to process packets.