plugins: support for source plugins
Allow a plugin to register itself as a source. This isn't that
much different than how current sources register, it just happens
a little later on during startup.

One "slot" is reserved for source plugins, but multiple plugins
implementing a source can be loaded.  The --source-plugin command
line option must be used to tell Suricata which plugin
to use.

This is still very much a work in progress, but can load the
PF_RING source as a plugin.