draft/output: Multi-threaded eve.json output
This is a single commit for POC purposes to gather feedback on the
approach and implementation.

Enable `outputs.eve-log.threaded` for lockless output to eve.json
files. A file is created for each thread -- the file is suffixed with
the thread id, e.g., `eve.json.7`. The entirety of `eve.json` files
(`eve.json.*`) must be aggregated and time-stitched into a single
(logical) eve.json container for upstream access.