track and log TCP flags in initial packets seen in each direction
Tested by recreating each of the TCP flow "flavours" that are
handled by yaf

Fix capture of initial TCP flags for each flow direction

Fixes an issue where the TCP flags for the 2nd packet from the server
were captured instead of the flags for the 1st packet.

Also improve handling of the scenario where the 1st observed packet for
a TCP flow is a RST or FIN and midstream-pickup is enabled via
configuration. The TCP flags are now captured in these situations.

Tested:
- [x] Checked initial TCP flags are correctly set in json records for all flow flavours
- [x] Ran all flow flavours against an AddressSanitizer build