iprep: Avoid incorrectly decrementing host use_cnt
If the packet already has the host_src or host_dst entry
set then the attached host entry is locked without increasing
the use_cnt. If the host is not already linked to the packet
then the use_cnt was incremented firstly in HostLookupHostFromHash
and then secondly in HostReference. We can avoid extra increment
and just unlock the host on function exit.