Improve midstream pickup of TCP flows at SYN/ACK stage
If midstream pickup is enabled for suricata, the first packet seen for
a flow is assumed to come from the client endpoint. This approach has a
50% chance of being correct. If the first packet to be seen for a TCP
flow is a SYN/ACK then it must have been sent by the server endpoint,
not the client. This change leverages that fact to improve the chances
of flows that are picked up midstream having the correct cardinality.

The FlowInit method now checks to see if the first packet is a TCP
SYN/ACK and, if it is, sets the source and destination addresses for
the flow accordingly. This subsequently means that the
PacketSwap/FlowSwap methods are no longer required, which in turn
*should* slightly reduce the computational overhead for affected flows.