enip: record direction in transaction and fixup detect
ENIP creates a transaction per message, so detection is run
twice on each transaction, one for each direction without
being away of the direction result in 2 alerts for each
transaction.

By recording the direction in the transaction, detection can now
skip matching if the detection is for the other direction.

Addresses redmine ticket:
https://redmine.openinfosecfoundation.org/issues/3886