etc/logrotate: update sample file to handle threaded eve
Move old log files to an "olddir". This will make it easier for
other tools consuming Eve files to operate, and not pick up rotated
log files.

Add threaded Eve log files to the filename match for rotation in a
that should not match existing rotated files such as "eve.json.1.1".
This is not strictly required when using olddir, but may prevent
issues when users only use a portion of our provided file as a
template.