detect/conf: Constrain max address length
This commit adds configuration support for a "max-address-length" that
governs the size of the maximum (unparsed) address value that can be
parsed.

When address values in the rule variables exceeds the default length
listed in the configuration value, use a higher value to accommodate the
size of the address values in use.

The default value is 8192 and can be overridden with the configuration
variable detect.max-address-size.