Makefile: use convenience library for C code
With the circular reference gone, compile all the C code into
a convenience library, and use this library when building
Suricata, as well as all the fuzz targets.

We initially tried this when the fuzz targets were added as it is
the proper way, but due to the cicular reference we used the
COMMON_SOURCES instead.

This puts us back in line with recommended automake usage when
multiple binaries share the same source files.