transform/md5: use new Rust md5 bindings
Convert the md5 transform from NSS to our new Rust bindings. This
also shows an example of how we might globally disable hashing.
Basically its just a global, g_disable_hashing that takes the
place of the HAVE_NSS guard. This would be a runtime configuration
value that is disabled by default, but would let users opt-in to
mode that is similar to building without NSS support.