datasets: fix permission issue
When suricata is started with the --user or --group options, there is
a permission issue with the dataset files. They are created by root and
end up being owned by root. As a consequence the dataset files can not be
saved at exit because the file are open after switch to lower privilege
and thus the action is done as the lower privileged user.

This patch updates the code to chown the dataset files after opening so
it belongs to the lower privileged user.